[RF]: MultiEchoData class
Beginning of class-style refactoring to ease up the global shenanigans going on
in `tedana.py`. Created a MultiEchoData class (`interfaces/data.py`) to hold
input multi-echo data, echo times, create masks, etc. Made stub classes for
other features that can be used to class-ify other aspects of the code. I still
think that the API should be function-based, but this might help behind the
scenes so that related information is stored together.