[STY] Mostly stylistic, a few enhancements...
Modified some code in utils.py to be more readable and to accept various input
types (e.g., lists of data files, arrays, etc.). In the process, changed a some
doc-strings, modified the print statements to be logging statements, and made a
few comments for identification of things that need to be changed to better
integrate surface files.