[DOC] Supercharged logging
Added a bunch of logging statements through the codebase, options to augment
(--debug) or stifle (--quiet) them, and some comments throughout. Also updated
`setup.py` so that once installed, `tedana` is callable from the command line
(it had erroneous pointers to meica).

Hoping I didn't change any of the results in the process of doing this...