Slight refactor.
NOTE: Now I know why the maps were different! In this refactor, I’m
cropping values in T2* limited and T2* full after masking, while before
we did it before. So any voxel with an adaptive mask value of 1 has a
value of 0 in the old T2* limited, but 1 in the new.